requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
bip-utils==2.9.3
bitarray==2.9.2
cbor2==5.6.4
cdp-sdk==0.0.8
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
coincurve==20.0.0
crcmod==1.7
cryptography==43.0.1
cytoolz==0.12.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
instructor==1.6.3
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mnemonic==0.21
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.2
openai==1.52.2
packaging==24.1
parsimonious==0.10.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rlp==4.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
swarm @ git+ssh://git@github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
tomli==2.0.2
toolz==0.12.1
tqdm==4.66.6
typer==0.12.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
